refactor: abstract away template-card into its own component
